Create checkpoint-07.swift
This was my own initial attempt without checking the solution. It functions and is pretty clean, but I missed nuance in the instructions and ended up creating INSTANCES of the Dog and Cat subclasses instead of further subclasses. Makes sense because then you could create instances of the breed subclass that are named and have individual attributes like color, weight, etc...